src/main/java/eu/tib/ts/controller/RatioController.java
package eu.tib.ts.controller;

import eu.tib.ts.controller.dto.RatioDto;
import eu.tib.ts.model.ontology.CharacteristicsType;
import eu.tib.ts.model.ontology.SimpleOntology;
import eu.tib.ts.service.RatioService;
import eu.tib.ts.utils.HttpUtils;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/api/ontology/ratio")
public class RatioController {
    private final RatioService ratioService;

    @Autowired
    public RatioController(RatioService ratioService) {
        this.ratioService = ratioService;
    }

    @PostMapping(value = "/{characteristics}", produces = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity<RatioDto> getRatio(
        @PathVariable("characteristics") CharacteristicsType characteristicsType,
        @RequestBody List<SimpleOntology> ontologies
    ) {
        RatioDto ratioDto = ratioService.getRatio(ontologies, characteristicsType);

        return HttpUtils.ok(ratioDto);
    }
}
